File Upload: Delete Image files when editing/deleting products

+++ controllers/admin.js
@@ -1,4 +1,5 @@
-const mongoose = require('mongoose'); // temp
+const mongoose = require('mongoose'); 
+const fileHelper = require('../util/file');
 const Product = require('../models/product');
 const { validationResult} = require('express-validator/check');
 
@@ -159,9 +160,9 @@ exports.postEditProduct = (req, res, next) => {
         product.price = updatedPrice;
         product.description = updatedDescription;
         if(image) {
+            fileHelper.deleteFile(product.imageUrl);
             product.imageUrl = image.path;
-        }
-        
+        }        
         return product.save().then(result => {
             console.log('UPDATED PRODUCT');
             res.redirect('/admin/products');
@@ -176,8 +177,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    //Product.findByIdAndRemove(prodId)
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product.findById(prodId)
+    .then(product => {
+        if(!product) {
+            return next(new Error('Product not found.'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({_id: prodId, userId: req.user._id})
+    })
     .then(() => {
         console.log('DESTROYED PRODUCT');
         res.redirect('/admin/products');
